+++ packages/messages-provider/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cloudoperators/juno-messages-provider
 
+## 0.2.20
+
+### Patch Changes
+
+- Updated dependencies [2125722]
+- Updated dependencies [42ad17d]
+  - @cloudoperators/juno-ui-components@5.8.0
+
 ## 0.2.19
 
 ### Patch Changes